+++ lib/dicey/cli/calculator_runner.rb
@@ -2,7 +2,7 @@
 
 require_relative "../die_foundry"
 
-Dir["../sum_frequency_calculators/*.rb", base: __dir__].each { require_relative _1 }
+Dir["../distribution_calculators/*.rb", base: __dir__].each { require_relative _1 }
 
 module Dicey
   module CLI
@@ -34,7 +34,7 @@ module Dicey
       end
 
       def calculator_selector
-        @calculator_selector ||= SumFrequencyCalculators::AutoSelector.new
+        @calculator_selector ||= DistributionCalculators::AutoSelector.new
       end
     end
   end