+++ spec/dicey/distribution_calculators/brute_force_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Dicey
-  RSpec.describe SumFrequencyCalculators::BruteForce do
+  RSpec.describe DistributionCalculators::BruteForce do
     subject(:result) { calculator.call(dice) }
 
     let(:calculator) { described_class.new }