+++ spec/dicey/distribution_calculators/kronecker_substitution_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Dicey
-  RSpec.describe SumFrequencyCalculators::KroneckerSubstitution do
+  RSpec.describe DistributionCalculators::KroneckerSubstitution do
     subject(:result) { calculator.call(dice) }
 
     let(:calculator) { described_class.new }